+++ side/src/main/java/com/blanc/side/SideApplication.java
@@ -1,45 +1,16 @@
 package com.blanc.side;
 
-import org.springframework.boot.web.embedded.tomcat.TomcatServletWebServerFactory;
-import org.springframework.boot.web.server.WebServer;
-import org.springframework.boot.web.servlet.server.ServletWebServerFactory;
-import org.springframework.context.annotation.Bean;
+import com.blanc.side.toby.bootApplication.MySpringbootApplication;
 import org.springframework.context.annotation.ComponentScan;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.context.support.AnnotationConfigWebApplicationContext;
-import org.springframework.web.servlet.DispatcherServlet;
 
-@Configuration
+
 @ComponentScan
 public class SideApplication {
 
-  @Bean
-  public ServletWebServerFactory serverFactory() {
-    return new TomcatServletWebServerFactory();
-  }
-
-  @Bean
-  public DispatcherServlet dispatcherServlet() {
-    return new DispatcherServlet();
-  }
-
   public static void main(String[] args) {
-    AnnotationConfigWebApplicationContext applicationContext = new AnnotationConfigWebApplicationContext() {
-      @Override
-      protected void onRefresh() {
-        super.onRefresh();
-
-        ServletWebServerFactory serverFactory = this.getBean(ServletWebServerFactory.class);
-        DispatcherServlet dispatcherServlet = this.getBean(DispatcherServlet.class);
-        WebServer webServer = serverFactory.getWebServer(
-            servletContext -> servletContext.addServlet("DispatcherServlet", dispatcherServlet).addMapping("/*"));
-        webServer.start();
-
-      }
-    };
-    applicationContext.register(SideApplication.class);
-    applicationContext.refresh();
+    MySpringbootApplication.run(SideApplication.class, args);
 
   }
+
 
 }
