+++ side/src/main/java/com/blanc/side/toby/bootApplication/MySpringbootApplication.java
@@ -1,0 +1,29 @@
+package com.blanc.side.toby.bootApplication;
+
+import org.springframework.boot.web.server.WebServer;
+import org.springframework.boot.web.servlet.server.ServletWebServerFactory;
+import org.springframework.context.annotation.ComponentScan;
+import org.springframework.web.context.support.AnnotationConfigWebApplicationContext;
+import org.springframework.web.servlet.DispatcherServlet;
+
+@ComponentScan
+public class MySpringbootApplication {
+
+  public static void run(Class<?> applicationClass, String... args) {
+    AnnotationConfigWebApplicationContext applicationContext = new AnnotationConfigWebApplicationContext() {
+      @Override
+      protected void onRefresh() {
+        super.onRefresh();
+
+        ServletWebServerFactory serverFactory = this.getBean(ServletWebServerFactory.class);
+        DispatcherServlet dispatcherServlet = this.getBean(DispatcherServlet.class);
+        WebServer webServer = serverFactory.getWebServer(
+            servletContext -> servletContext.addServlet("DispatcherServlet", dispatcherServlet).addMapping("/*"));
+        webServer.start();
+
+      }
+    };
+    applicationContext.register(applicationClass);
+    applicationContext.refresh();
+  }
+}
