+++ side/src/main/java/com/blanc/side/toby/config/config.java
@@ -1,0 +1,21 @@
+package com.blanc.side.toby.config;
+
+import org.springframework.boot.web.embedded.tomcat.TomcatServletWebServerFactory;
+import org.springframework.boot.web.servlet.server.ServletWebServerFactory;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.servlet.DispatcherServlet;
+
+@Configuration
+public class config {
+
+  @Bean
+  public ServletWebServerFactory serverFactory() {
+    return new TomcatServletWebServerFactory();
+  }
+
+  @Bean
+  public DispatcherServlet dispatcherServlet() {
+    return new DispatcherServlet();
+  }
+}
